flowc: nim: almost all functions from 'dynamic' are implemented

+++ tools/flowc/backends/nim/natives/ds/array/foldi.nim
@@ -2,5 +2,5 @@
 proc foldi*[T, S](xs: seq[T], init: S, fn: proc(idx: int32, acc: S, v: T): S): S =
   var ini = init
   for i in 0..xs.len - 1:
-    ini = fn(i, ini, xs[i])
+    ini = fn(int32(i), ini, xs[i])
   return ini
